bench: add benchmarks
