+++ Cargo.toml
@@ -5,14 +5,15 @@ edition = "2024"
 
 [dev-dependencies]
 criterion = { version = "0.5", features = ["html_reports"] }
+regex = "1.11.1"
 
 [[bench]]
 name = "thompson_benchmark"
 harness = false
 
-[[bench]]
-name = "glushkov_benchmark"
-harness = false
+# [[bench]]
+# name = "glushkov_benchmark"
+# harness = false
 
 [[bench]]
 name = "rust_regex_benchmark"
